+++ Controllers/VentasController.cs
@@ -21,26 +21,37 @@ namespace VentaProductos.Controllers
         }
 
         // GET: api/Ventas
-        // [HttpGet]
-        // public async Task<ActionResult<IEnumerable<Venta>>> GetVentas()
-        // {
-        //     return await _context.Ventas.ToListAsync();
-        // }
-
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Venta>>> GetVentas()
         {
-            return await _context.Ventas.Include(v => v.DetalleVenta).ToListAsync();
+            var ventas = await _context.Ventas.Include(x => x.Cliente).ToListAsync();
+
+            return ventas;
         }
 
-
         // GET: api/Ventas/5
+        // [HttpGet("{id}")]
+        // public async Task<ActionResult<Venta>> GetVenta(int id)
+        // {
+        //     var venta = await _context.Ventas
+        //                                 .Include
+        //     FindAsync(id);
+
+        //     if (venta == null)
+        //     {
+        //         return NotFound();
+        //     }
+
+        //     return venta;
+        // }
+
+
         [HttpGet("{id}")]
         public async Task<ActionResult<Venta>> GetVenta(int id)
         {
             var venta = await _context.Ventas
-                                    .Include(v => v.DetalleVenta)
-                                    .FirstOrDefaultAsync(v => v.Id == id);
+                                        .Include(v => v.Cliente) 
+                                        .FirstOrDefaultAsync(v => v.Id == id);
 
             if (venta == null)
             {
@@ -49,6 +60,7 @@ namespace VentaProductos.Controllers
 
             return venta;
         }
+
 
         // PUT: api/Ventas/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
@@ -78,51 +90,28 @@ namespace VentaProductos.Controllers
                 }
             }
 
-            return NoContent();
+            return Ok(venta);
         }
-
-        // [HttpPut("{id}/finalizar")]
-        // public async Task<IActionResult> FinalizarVenta (int id)
-        // {
-        //     var venta = await _context.Ventas.FindAsync(id);
-        //     if (venta == null) return NotFound();
-
-        //     if (venta.Finalizada)
-        //     {
-        //         return BadRequest ("La venta ya esta finalizada");
-        //     }
-
-        //     venta.Finalizada = true;
-        //     _context.Entry(venta).State = EntityState.Modified;
-        //     await _context.SaveChangesAsync();
-
-        //     return NoContent();
-        // }
 
         // POST: api/Ventas
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
-        public async Task<ActionResult<Venta>> PostVenta(Venta venta, List<int> productosIds)
+        public async Task<ActionResult<Venta>> PostVenta(Venta venta)
         {
-
-            // Verificar si el cliente tiene una venta sin finalizar
-            var ventaActiva = await _context.Ventas
-                                    .Where(v => v.IdCliente == venta.IdCliente && !v.Finalizada)
-                                    .FirstOrDefaultAsync();
-            if (ventaActiva != null)
+            var cliente = await _context.Clientes
+                                        .Include(c => c.Ventas)
+                                        .FirstOrDefaultAsync(c => c.Id == venta.IdCliente);
+                
+            if (cliente == null)
             {
-                return BadRequest("El cliente ya tiene una venta activa. Debe finalizar la venta antes de crear una nueva.");
+                return BadRequest("El cliente no existe");
             }
 
-        //crear la venta y sus detalles
+            if (cliente.Ventas != null && cliente.Ventas.Any(venta => venta.Finalizada != true))
+            {
+                return BadRequest("El cliente tiene pendiente una venta");
+            }
             _context.Ventas.Add(venta);
-            await _context.SaveChangesAsync();
-
-
-            foreach (var ProductoId in productosIds)
-            {
-                _context.DetalleVentas.Add(new DetalleVenta { IdProducto = ProductoId, IdVenta = venta.Id });
-            }
             await _context.SaveChangesAsync();
 
             return CreatedAtAction("GetVenta", new { id = venta.Id }, venta);
@@ -137,7 +126,13 @@ namespace VentaProductos.Controllers
             {
                 return NotFound();
             }
-         _context.Ventas.Remove(venta);
+
+            if (!venta.Finalizada)
+            {
+                return BadRequest("La venta no puede ser eliminada porque no esta finalizada");
+            }
+
+            _context.Ventas.Remove(venta);
             await _context.SaveChangesAsync();
 
             return NoContent();
